feature: show icon when downloading tiles for offline use
It can be difficult to tell when the app is in progress of downloading map tiles
for offline use, especially when hitting a slow server like USGS or on a slow
network. Adds a download circle icon to indicate download progress to the
user. Also helpfully informs when a zoom level is outside the range.